lls-dagster-pipeline/pyproject.toml
[project]
name = "lls_dagster_pipeline"
requires-python = ">=3.11,<3.12"
version = "0.1.0"
dependencies = [
    "dagster==1.12.2",
    "fastapi[standard]>=0.121.2",
    "holidays>=0.85",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "mlflow>=3.6.0",
    "numpy>=1.23.5, <2.0.0",
    "pandas>=2.3.3",
    "shapely>=2.1.2",
    "tensorflow-macos==2.15.0",
    "tensorflow-metal==1.1.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "lls_dagster_pipeline"
registry_modules = [
    "lls_dagster_pipeline.components.*",
]


